Change users dir to use test runner
