+++ users/test_runner.py
@@ -1,0 +1,33 @@
+"""Unit test runner for users service.
+
+Test runner that runs several different test modules.
+To see the tests, look at the sibling test_*.py files."""
+
+# Authors: mukobi
+# Copyright 2019 The Knative Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import unittest
+
+import test_authorization
+import test_user_upsertion
+
+
+if __name__ == "__main__":
+    loader = unittest.TestLoader()
+    suite = unittest.TestSuite()
+    suite.addTest(loader.loadTestsFromModule(test_authorization))
+    suite.addTest(loader.loadTestsFromModule(test_user_upsertion))
+
+    unittest.TextTestRunner().run(suite)
