+++ users/test_user_upsertion.py
@@ -1,4 +1,4 @@
-"""Unit tests for users service."""
+"""Unit tests for users service user upsertion functions."""
 
 # Authors: mukobi
 # Copyright 2019 The Knative Authors
@@ -19,52 +19,9 @@ import unittest
 import mongomock
 import app
 
-AUTHORIZED_USER = "authorized-user"
-NON_AUTHORIZED_USER = "non-authorized-user"
-MALFORMATTED_IN_DB_USER = "this-user-has-no-'is_organizer'_db_field"
-MISSING_USER = "not-a-user-in-the-database"
-
 USER_ID = "valid-user-id"
 USER_NAME = "Valid User Name"
 ADDITIONAL_INFORMATION = "I'm not found in other documents"
-
-
-class TestAuthorization(unittest.TestCase):
-    """Test /authorization endpoint of users service."""
-
-    def setUp(self):
-        """Seed mock DB for testing"""
-        self.mock_collection = mongomock.MongoClient().db.collection
-        mock_data = [
-            {"user_id": AUTHORIZED_USER,
-             "name": AUTHORIZED_USER,
-             "is_organizer": True},
-            {"user_id": NON_AUTHORIZED_USER,
-             "name": NON_AUTHORIZED_USER,
-             "is_organizer": False},
-            {"user_id": MALFORMATTED_IN_DB_USER}
-        ]
-        self.mock_collection.insert_many(mock_data)
-
-    def test_authorized_user_is_authorized(self):
-        """An authorized user should receive authorized privileges."""
-        self.assertTrue(app.find_authorization_in_db(
-            AUTHORIZED_USER, self.mock_collection))
-
-    def test_non_authorized_user_not_authorized(self):
-        """A non-authorized user should not receive authorized privileges."""
-        self.assertFalse(app.find_authorization_in_db(
-            NON_AUTHORIZED_USER, self.mock_collection))
-
-    def test_malformatted_user_none_authorization(self):
-        """An incorrect db schema should not receive authorized privileges."""
-        self.assertFalse(app.find_authorization_in_db(
-            MISSING_USER, self.mock_collection))
-
-    def test_unkown_user_not_authorized(self):
-        """A user not in the db should not receive authorized privileges."""
-        self.assertFalse(app.find_authorization_in_db(
-            MALFORMATTED_IN_DB_USER, self.mock_collection))
 
 
 class TestUserUpsertion(unittest.TestCase):
